server/notifier/requirements.txt
paho-mqtt
python-telegram-bot